Break ConfigSelectModal into separate file
