+++ scss/linodes/components/ConfigSelectModal.scss
@@ -1,0 +1,15 @@
+.ConfigSelectModal {
+    &-configs {
+        border: $border;
+        height: 200px;
+        overflow-y: scroll;
+        overflow-x: hidden;
+        padding: $input-padding;
+
+        label {
+            display: block;
+            white-space: nowrap;
+        }
+    }
+}
+
