+++ scss/linodes/components/StatusDropdown.scss
@@ -1,19 +1,3 @@
-.LinodesComponentsStatusDropdown {
-    &-configs {
-        border: $border;
-        height: 200px;
-        overflow-y: scroll;
-        overflow-x: hidden;
-        padding: $input-padding;
-
-        label {
-            display: block;
-            padding-bottom: $input-padding / 2;
-            white-space: nowrap;
-        }
-    }
-}
-
 .StatusDropdown {
     @extend .btn-group;
