+++ scss/linodes/components/index.scss
@@ -1,3 +1,4 @@
+@import 'ConfigSelectModal';
 @import 'Distribution';
 @import 'Distributions';
 @import 'StatusDropdown';
